requirements.txt
# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
torchmetrics>=0.11.0

# Image processing
opencv-python>=4.5.5
Pillow>=9.0.0
scikit-image>=0.19.0

# Scientific computing
numpy>=1.21
scipy>=1.7.3

# Utilities
matplotlib>=3.5
PyYAML>=6.0
tqdm>=4.64
loguru>=0.6.0

# Logging and visualization
tensorboard>=2.8
seaborn>=0.11.2

# Transformers and attention mechanisms
transformers>=4.26

# Compatibility & CLI
click>=8.0.3

# HuggingFace Dataset (if using external data)
datasets>=2.4.0

# Optional for visualization/debug
ipykernel
jupyter

# Version compatibility
python>=3.8